fix: position designations on cycle positions
