+++ talentmap_api/position/models.py
@@ -226,10 +226,6 @@ class PositionBidStatistics(StaticRepresentationModel):
     has_handshake_offered = models.BooleanField(default=False)
     has_handshake_accepted = models.BooleanField(default=False)
 
-    is_urgent_vacancy = models.BooleanField(default=False)
-    is_volunteer = models.BooleanField(default=False)
-    is_hard_to_fill = models.BooleanField(default=False)
-
     def update_statistics(self):
         bidcycle_bids = self.position.bids.filter(bidcycle=self.position.bidcycle)
         self.total_bids = bidcycle_bids.count()
